chore: move main to the cmd
